Heart Failure Case Prediction by Machine Learning
An application of machine learning to perform retrospective analysis on hospital admittance for heart failure cases. After treating a patient, hospitals make a decision on whether to discharge a patient or not if they predict symptoms will re-emerge. This decision if incorrect will cost money and resources as re-admitting patients or keeping a healthy patients in the hospital is costly. This application uses machine learning in order to better predict whether to keep or discharge patients using information about the patient such as their age or past medical history in conjunction with retrospective data on whether the patient needed to be re-admitted or not. Additionally, this application outlines the largest factors that lead to heart failure.